[backend - 박규민] 2024.11.20- 유저아이디와 아이템아이디 모두를 입력해야 장바구니가 삭제가능하도록 api수정

+++ Book/src/main/java/com/book/cart/CartService.java
@@ -25,8 +25,11 @@ public class CartService {
         return cartRepository.findByUserid(userid);
     }
 
-    // 장바구니에서 책 삭제
-    public void removeBookFromCart(Long itemId) {
-        cartRepository.deleteById(itemId);
+    // 장바구니에서 책 삭제방법
+    public void removeBookFromCart(String userId, Long itemId) {
+        CartItem cartItem = cartRepository.findById(itemId).orElse(null);
+        if (cartItem != null && cartItem.getUserid().equals(userId)) {
+            cartRepository.deleteById(itemId);
+        }
     }
 }
